refactor: split wrangler related logic
